requirements.txt
asgiref==3.8.1
Django==5.1.1
sqlparse==0.5.1
tzdata==2024.2
psycopg2-binary==2.9.9
django-allauth==65.0.2
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
django-allauth[socialaccount]
sendgrid==6.11.0
django-environ==0.11.2
pillow==10.4.0
django-debug-toolbar==4.4.6
whitenoise==6.7.0
gunicorn==22.0.0